Hai Anh -  commit search ebook. Not done.

+++ src/main/java/com/higgsup/intern/ebshop/jdbc/dao/EbookDAO.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface EbookDAO {
     Ebook findById(Long id);
-    List<Ebook> find(String name, Long authorId, Long publisherId, Long priceFrom, Long priceTo, String isbn);
+   List<Ebook> find(String name, Long authorId, Long publisherId, Double priceFrom, Double priceTo, String isbn);
     Ebook findByIsbn(String isbn);
     List<EbookOrderDTO> findTop10BestSellerEbooks();
     Publisher getPublisherByEbookId(Long id);
